requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Computer Vision & ML
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200  # YOLOv8
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
pillow==10.1.0

# OCR
paddleocr==2.7.0.3
easyocr==1.7.1
pytesseract==0.3.10

# Deep Learning
tensorflow==2.15.0
keras==2.15.0
onnx==1.15.0
onnxruntime==1.16.3

# Image Processing
albumentations==1.3.1
scikit-image==0.22.0
imageio==2.33.0

# Data Processing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Storage
boto3==1.29.7  # AWS S3
google-cloud-storage==2.13.0  # GCS

# API & Web
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0
websockets==12.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
locust==2.19.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
colorlog==6.8.0

# Document Generation
reportlab==4.0.7
fpdf2==2.7.6
python-docx==1.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pylint==3.0.3

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
